refactor: move walkShallow to portlist
